devtools/client/webide/test/test_basic.html

<!DOCTYPE html>

<html>

  <head>
    <meta charset="utf8">
    <title></title>

    <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript" src="chrome://mochikit/content/chrome-harness.js"></script>
    <script type="application/javascript" src="head.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  </head>

  <body>

    <script type="application/javascript">
      window.onload = function() {
        SimpleTest.waitForExplicitFinish();

        Task.spawn(function* () {
            let win = yield openWebIDE();

            const {gDevToolsBrowser} = require("devtools/client/framework/devtools-browser");
            yield gDevToolsBrowser.isWebIDEInitialized.promise;
            ok(true, "WebIDE was initialized");

            ok(win, "Found a window");
            ok(win.AppManager, "App Manager accessible");
            let appmgr = win.AppManager;
            ok(appmgr.connection, "App Manager connection ready");
            ok(appmgr.runtimeList, "Runtime list ready");

            // test error reporting
            let nbox = win.document.querySelector("#notificationbox");
            let notification =  nbox.getNotificationWithValue("webide:errornotification");
            ok(!notification, "No notification yet");
            let deferred = new Promise((resolve, reject) => {
              nextTick().then(() => {
                reject("BOOM!");
              });
            });
            try {
              yield win.UI.busyUntil(deferred, "xx");
            } catch(e) {/* This *will* fail */}
            notification =  nbox.getNotificationWithValue("webide:errornotification");
            ok(notification, "Error has been reported");

            yield closeWebIDE(win);

            SimpleTest.finish();
        });
      }
    </script>
  </body>
</html>
